Renames function to ga_async_snippet and loads in theme header.
The Google Analytics scripts in this plugin are using Mathias Bynens's
optimizations <http://mathiasbynens.be/notes/async-analytics-snippet>. Bynens
points out in his post that Google recommends putting the Analytics script as
high in the document as possible to take advantage of HTML5's asynchronous
script loading. Per that recommendation, this change adds the GA script to the
theme header. It also renames the script function to remove any implied ties to a
specific plugin hook or theme location.